moving alerts to device, fixing remove trigger naming conventions
