--- src/DataBoxEdge/DataBoxEdge/Common/Cmdlets/Alerts/DataBoxEdgeDeviceGetCmdletBase.cs
@@ -1,152 +1,0 @@
-﻿// ----------------------------------------------------------------------------------
-//
-// Copyright Microsoft Corporation
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-// http://www.apache.org/licenses/LICENSE-2.0
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-// ----------------------------------------------------------------------------------
-
-using System.Collections.Generic;
-using System.Linq;
-using System.Management.Automation;
-using Microsoft.Azure.Commands.ResourceManager.Common.ArgumentCompleters;
-using Microsoft.Azure.Management.EdgeGateway;
-using Microsoft.Rest.Azure;
-using Microsoft.WindowsAzure.Commands.Utilities.Common;
-using ResourceModel = Microsoft.Azure.Management.EdgeGateway.Models.Alert;
-using PSResourceModel = Microsoft.Azure.PowerShell.Cmdlets.DataBoxEdge.Models.PSDataBoxEdgeAlert;
-using PSTopLevelResourceObject = Microsoft.Azure.PowerShell.Cmdlets.DataBoxEdge.Models.PSDataBoxEdgeDevice;
-
-namespace Microsoft.Azure.PowerShell.Cmdlets.DataBoxEdge.Common.Cmdlets.Alerts
-{
-    [Cmdlet(VerbsCommon.Get, Constants.Alert, DefaultParameterSetName = ListByParameterSet
-     ),
-     OutputType(typeof(PSResourceModel)),
-    ]
-    public class DataBoxEdgeAlertGetCmdletBase : AzureDataBoxEdgeCmdletBase
-    {
-        private const string ListByParameterSet = "ListByParameterSet";
-        private const string GetByResourceIdParameterSet = "GetByResourceIdParameterSet";
-        private const string GetByDeviceObjectParameterSet = "GetByDeviceObjectParameterSet";
-        private const string GetByNameParameterSet = "GetByNameParameterSet";
-
-        [Parameter(Mandatory = true,
-            ParameterSetName = GetByResourceIdParameterSet,
-            HelpMessage = Constants.ResourceIdHelpMessage)]
-        [ValidateNotNullOrEmpty]
-        public string ResourceId { get; set; }
-
-        [Parameter(Mandatory = true,
-            ParameterSetName = ListByParameterSet,
-            HelpMessage = Constants.ResourceGroupNameHelpMessage,
-            Position = 0)]
-        [Parameter(Mandatory = true,
-            ParameterSetName = GetByNameParameterSet,
-            HelpMessage = Constants.ResourceGroupNameHelpMessage,
-            Position = 0)]
-        [ValidateNotNullOrEmpty]
-        [ResourceGroupCompleter]
-        public string ResourceGroupName { get; set; }
-
-        [Parameter(Mandatory = true,
-            ParameterSetName = ListByParameterSet,
-            HelpMessage = Constants.ResourceGroupNameHelpMessage,
-            Position = 0)]
-        [Parameter(Mandatory = true,
-            ParameterSetName = GetByNameParameterSet,
-            HelpMessage = Constants.DeviceNameHelpMessage,
-            Position = 1)]
-        [ResourceNameCompleter("Microsoft.DataBoxEdge/dataBoxEdgeDevices", nameof(ResourceGroupName))]
-        [ValidateNotNullOrEmpty]
-        public string DeviceName { get; set; }
-
-        [Parameter(Mandatory = true,
-            ParameterSetName = GetByNameParameterSet,
-            HelpMessage = Constants.NameHelpMessage,
-            Position = 2)]
-        [ValidateNotNullOrEmpty]
-        public string Name { get; set; }
-
-
-        [Parameter(Mandatory = true,
-            ParameterSetName = GetByDeviceObjectParameterSet,
-            HelpMessage = Constants.PsDeviceObjectHelpMessage,
-            Position = 2)]
-        [ValidateNotNullOrEmpty]
-        public PSTopLevelResourceObject DeviceObject { get; set; }
-
-
-        private ResourceModel GetResourceModel()
-        {
-            return AlertsOperationsExtensions.Get(
-                this.DataBoxEdgeManagementClient.Alerts,
-                this.DeviceName,
-                this.Name,
-                this.ResourceGroupName);
-        }
-
-        private IPage<ResourceModel> ListResourceModel()
-        {
-            return AlertsOperationsExtensions.ListByDataBoxEdgeDevice(
-                this.DataBoxEdgeManagementClient.Alerts,
-                this.DeviceName,
-                this.ResourceGroupName);
-        }
-
-        private IPage<ResourceModel> ListResourceModel(string nextPageLink)
-        {
-            return AlertsOperationsExtensions.ListByDataBoxEdgeDeviceNext(
-                this.DataBoxEdgeManagementClient.Alerts,
-                nextPageLink);
-        }
-
-        private List<PSResourceModel> GetByResourceName()
-        {
-            var resourceModel = GetResourceModel();
-            return new List<PSResourceModel>() {new PSResourceModel(resourceModel)};
-        }
-
-        private List<PSResourceModel> ListForEverything()
-        {
-            if (!string.IsNullOrEmpty(this.Name))
-            {
-                return GetByResourceName();
-            }
-
-            var resourceModels = ListResourceModel();
-            var paginatedResult = new List<ResourceModel>(resourceModels);
-            while (!string.IsNullOrEmpty(resourceModels.NextPageLink))
-            {
-                resourceModels = ListResourceModel(resourceModels.NextPageLink);
-                paginatedResult.AddRange(resourceModels);
-            }
-
-            return paginatedResult.Select(t => new PSResourceModel(t)).ToList();
-        }
-
-        public override void ExecuteCmdlet()
-        {
-            if (this.IsParameterBound(c => c.ResourceId))
-            {
-                var resourceIdentifier = new DataBoxEdgeResourceIdentifier(this.ResourceId);
-                this.ResourceGroupName = resourceIdentifier.ResourceGroupName;
-                this.DeviceName = resourceIdentifier.DeviceName;
-                this.Name = resourceIdentifier.Name;
-            }
-
-            if (this.IsParameterBound(c => c.DeviceObject))
-            {
-                this.ResourceGroupName = this.DeviceObject.ResourceGroupName;
-                this.DeviceName = this.DeviceObject.Name;
-            }
-
-            WriteObject(ListForEverything(), true);
-        }
-    }
-}
