+++ src/DataBoxEdge/DataBoxEdge/Common/Cmdlets/Devices/DataBoxEdgeDeviceGetCmdletBase.cs
@@ -41,12 +41,16 @@ namespace Microsoft.Azure.PowerShell.Cmdlets.DataBoxEdge.Common.Cmdlets.Devices
         private const string ListByParameterSet = "ListByParameterSet";
         private const string GetByResourceIdParameterSet = "GetByResourceIdParameterSet";
         private const string GetByNameParameterSet = "GetByNameParameterSet";
+
         private const string GetExtendedInfoParameterSet = "GetExtendedInfoParameterSet";
         private const string GetNetworkSettingParameterSet = "GetNetworkSettingParameterSet";
         private const string GetSummaryUpdateParameterSet = "GetSummaryUpdateParameterSet";
+        private const string GetAlertParameterSet = "GetAlertParameterSet";
+
         private const string GetExtendedInfoByResourceIdParameterSet = "GetExtendedInfoByResourceIdParameterSet";
         private const string GetNetworkSettingByResourceIdParameterSet = "GetNetworkSettingByResourceIdParameterSet";
         private const string GetSummaryUpdateByResourceIdParameterSet = "GetSummaryUpdateByResourceIdParameterSet";
+        private const string GetAlertByResourceIdParameterSet = "GetAlertByResourceIdParameterSet";
 
         [Parameter(Mandatory = true,
             ParameterSetName = GetByResourceIdParameterSet,
@@ -59,6 +63,9 @@ namespace Microsoft.Azure.PowerShell.Cmdlets.DataBoxEdge.Common.Cmdlets.Devices
             HelpMessage = Constants.ResourceIdHelpMessage)]
         [Parameter(Mandatory = true,
             ParameterSetName = GetSummaryUpdateByResourceIdParameterSet,
+            HelpMessage = Constants.ResourceIdHelpMessage)]
+        [Parameter(Mandatory = true,
+            ParameterSetName = GetAlertByResourceIdParameterSet,
             HelpMessage = Constants.ResourceIdHelpMessage)]
         [ValidateNotNullOrEmpty]
         public string ResourceId { get; set; }
@@ -83,6 +90,10 @@ namespace Microsoft.Azure.PowerShell.Cmdlets.DataBoxEdge.Common.Cmdlets.Devices
             ParameterSetName = GetExtendedInfoParameterSet,
             HelpMessage = Constants.ResourceGroupNameHelpMessage,
             Position = 0)]
+        [Parameter(Mandatory = true,
+            ParameterSetName = GetAlertParameterSet,
+            HelpMessage = Constants.ResourceGroupNameHelpMessage,
+            Position = 0)]
         [ValidateNotNullOrEmpty]
         [ResourceGroupCompleter]
         public string ResourceGroupName { get; set; }
@@ -103,6 +114,10 @@ namespace Microsoft.Azure.PowerShell.Cmdlets.DataBoxEdge.Common.Cmdlets.Devices
             ParameterSetName = GetExtendedInfoParameterSet,
             HelpMessage = Constants.ResourceGroupNameHelpMessage,
             Position = 1)]
+        [Parameter(Mandatory = true,
+            ParameterSetName = GetAlertParameterSet,
+            HelpMessage = Constants.ResourceGroupNameHelpMessage,
+            Position = 1)]
         [ResourceNameCompleter("Microsoft.DataBoxEdge/dataBoxEdgeDevices", nameof(ResourceGroupName))]
         [ValidateNotNullOrEmpty]
         public string Name { get; set; }
@@ -110,17 +125,35 @@ namespace Microsoft.Azure.PowerShell.Cmdlets.DataBoxEdge.Common.Cmdlets.Devices
         [Parameter(Mandatory = true,
             ParameterSetName = GetExtendedInfoParameterSet,
             HelpMessage = Constants.ResourceGroupNameHelpMessage)]
+        [Parameter(Mandatory = true,
+            ParameterSetName = GetExtendedInfoByResourceIdParameterSet,
+            HelpMessage = Constants.ResourceGroupNameHelpMessage)]
         [ValidateNotNullOrEmpty]
         public SwitchParameter ExtendedInfo { get; set; }
 
         [Parameter(Mandatory = true,
             ParameterSetName = GetNetworkSettingParameterSet,
             HelpMessage = Constants.ResourceGroupNameHelpMessage)]
+        [Parameter(Mandatory = true,
+            ParameterSetName = GetNetworkSettingByResourceIdParameterSet,
+            HelpMessage = Constants.ResourceGroupNameHelpMessage)]
         [ValidateNotNullOrEmpty]
         public SwitchParameter NetworkSetting { get; set; }
 
         [Parameter(Mandatory = true,
+            ParameterSetName = GetAlertParameterSet,
+            HelpMessage = HelpMessageDevice.AlertHelpMessage)]
+        [Parameter(Mandatory = true,
+            ParameterSetName = GetAlertByResourceIdParameterSet,
+            HelpMessage = HelpMessageDevice.AlertHelpMessage)]
+        [ValidateNotNullOrEmpty]
+        public SwitchParameter Alert { get; set; }
+
+        [Parameter(Mandatory = true,
             ParameterSetName = GetSummaryUpdateParameterSet,
+            HelpMessage = Constants.ResourceGroupNameHelpMessage)]
+        [Parameter(Mandatory = true,
+            ParameterSetName = GetSummaryUpdateByResourceIdParameterSet,
             HelpMessage = Constants.ResourceGroupNameHelpMessage)]
         [ValidateNotNullOrEmpty]
         public SwitchParameter UpdateSummary { get; set; }
@@ -215,6 +248,14 @@ namespace Microsoft.Azure.PowerShell.Cmdlets.DataBoxEdge.Common.Cmdlets.Devices
                 this.ResourceGroupName));
         }
 
+
+        private List<PSDataBoxEdgeAlert> GetAlert()
+        {
+            var alerts = new DataBoxEdgeAlert(this.DataBoxEdgeManagementClient, this.ResourceGroupName, this.Name)
+                .Get();
+            return alerts;
+        }
+
         public override void ExecuteCmdlet()
         {
             var results = new List<PSResourceModel>();
@@ -238,6 +279,10 @@ namespace Microsoft.Azure.PowerShell.Cmdlets.DataBoxEdge.Common.Cmdlets.Devices
             {
                 var info = GetUpdatedSummary();
                 WriteObject(info, enumerateCollection: true);
+            }
+            else if (this.Alert.IsPresent)
+            {
+                WriteObject(GetAlert(), enumerateCollection: true);
             }
             else
             {
