+++ src/DataBoxEdge/DataBoxEdge/Common/Cmdlets/Devices/HelpMessageDevice.cs
@@ -23,5 +23,6 @@ namespace Microsoft.Azure.PowerShell.Cmdlets.DataBoxEdge.Common.Cmdlets.Devices
         internal const string ScanUpdateHelpMessage = "Scans for updates on a data box edge/gateway device.";
         internal const string FetchUpdateUpdateHelpMessage = "Downloads the updates on a data box edge/gateway device";
         internal const string InstallUpdateHelpMessage = "Installs the updates on the data box edge/gateway device";
+        internal const string AlertHelpMessage = "Fetch the alerts on the data box edge/gateway device";
     }
 }
