+++ src/DataBoxEdge/DataBoxEdge/help/Az.DataBoxEdge.md
@@ -11,9 +11,6 @@ Locale:
 Module to create a device and perform operations on DataBoxEdge devices
 
 ## Az.DataBoxEdge Cmdlets
-### [Get-AzDataBoxEdgeAlert](Get-AzDataBoxEdgeAlert.md)
-Get the alerts for the device
-
 ### [Get-AzDataBoxEdgeBandwidthSchedule](Get-AzDataBoxEdgeBandwidthSchedule.md)
 Gets the information about Bandwidth schedules
 
@@ -31,6 +28,9 @@ Gets the creted shares for this device
 
 ### [Get-AzDataBoxEdgeStorageAccountCredential](Get-AzDataBoxEdgeStorageAccountCredential.md)
 Get the Storage Account credential corresponding to device and storage account
+
+### [Get-AzDataBoxEdgeTrigger](Get-AzDataBoxEdgeTrigger.md)
+Get the Triggers configured on the device
 
 ### [Get-AzDataBoxEdgeUser](Get-AzDataBoxEdgeUser.md)
 Get the created users  for this device
@@ -53,6 +53,9 @@ Creates a new share in the device
 ### [New-AzDataBoxEdgeStorageAccountCredential](New-AzDataBoxEdgeStorageAccountCredential.md)
 Create new storage account credential object
 
+### [New-AzDataBoxEdgeTrigger](New-AzDataBoxEdgeTrigger.md)
+To Configure Triggers 
+
 ### [New-AzDataBoxEdgeUser](New-AzDataBoxEdgeUser.md)
 Creates a new user for the device
 
@@ -70,6 +73,9 @@ Removes the assosciated Role for the device
 
 ### [Remove-AzDataBoxEdgeStorageAccountCredential](Remove-AzDataBoxEdgeStorageAccountCredential.md)
 Removes a storage account credentail object for the device
+
+### [Remove-AzDataBoxEdgeTrigger](Remove-AzDataBoxEdgeTrigger.md)
+{{ Fill in the Synopsis }}
 
 ### [Remove-AzDataBoxEdgeUser](Remove-AzDataBoxEdgeUser.md)
 Removes the user
