--- src/DataBoxEdge/DataBoxEdge/help/Get-AzDataBoxEdgeAlert.md
@@ -1,157 +1,0 @@
----
-external help file: Microsoft.Azure.PowerShell.Cmdlets.DataBoxEdge.dll-Help.xml
-Module Name: Az.DataBoxEdge
-online version:
-schema: 2.0.0
----
-
-# Get-AzDataBoxEdgeAlert
-
-## SYNOPSIS
-Get the alerts for the device
-
-## SYNTAX
-
-### ListByParameterSet (Default)
-```
-Get-AzDataBoxEdgeAlert [-ResourceGroupName] <String> [-DeviceName] <String>
- [-DefaultProfile <IAzureContextContainer>] [<CommonParameters>]
-```
-
-### GetByResourceIdParameterSet
-```
-Get-AzDataBoxEdgeAlert -ResourceId <String> [-DefaultProfile <IAzureContextContainer>] [<CommonParameters>]
-```
-
-### GetByNameParameterSet
-```
-Get-AzDataBoxEdgeAlert [-ResourceGroupName] <String> [-DeviceName] <String> [-Name] <String>
- [-DefaultProfile <IAzureContextContainer>] [<CommonParameters>]
-```
-
-### GetByDeviceObjectParameterSet
-```
-Get-AzDataBoxEdgeAlert [-DeviceObject] <PSDataBoxEdgeDevice> [-DefaultProfile <IAzureContextContainer>]
- [<CommonParameters>]
-```
-
-## DESCRIPTION
-The **Get-AzDataBoxEdgeAlert** cmdlet gets alerts for the device. If you specify the Name of the Alert along with the resource group name and Device name, this cmdlet gets information about that specific Alert
-
-## EXAMPLES
-
-### Example 1
-```powershell
-PS C:\> Get-AzDataBoxEdgeAlert -ResourceGroupName psrgpfortest -DeviceName psdataboxedgedevice
-
-Title                            Severity AppearedDateTime
------                            -------- ----------------
-Lost heartbeat from your device. Critical 18-Oct-19 5:55:26 AM
-```
-
-## PARAMETERS
-
-### -DefaultProfile
-The credentials, account, tenant, and subscription used for communication with Azure.
-
-```yaml
-Type: Microsoft.Azure.Commands.Common.Authentication.Abstractions.Core.IAzureContextContainer
-Parameter Sets: (All)
-Aliases: AzContext, AzureRmContext, AzureCredential
-
-Required: False
-Position: Named
-Default value: None
-Accept pipeline input: False
-Accept wildcard characters: False
-```
-
-### -DeviceName
-Resource Group Name
-
-```yaml
-Type: System.String
-Parameter Sets: ListByParameterSet, GetByNameParameterSet
-Aliases:
-
-Required: True
-Position: 0
-Default value: None
-Accept pipeline input: False
-Accept wildcard characters: False
-```
-
-### -DeviceObject
-Please provide corresponding device object
-
-```yaml
-Type: Microsoft.Azure.PowerShell.Cmdlets.DataBoxEdge.Models.PSDataBoxEdgeDevice
-Parameter Sets: GetByDeviceObjectParameterSet
-Aliases:
-
-Required: True
-Position: 2
-Default value: None
-Accept pipeline input: False
-Accept wildcard characters: False
-```
-
-### -Name
-Resource Name
-
-```yaml
-Type: System.String
-Parameter Sets: GetByNameParameterSet
-Aliases:
-
-Required: True
-Position: 2
-Default value: None
-Accept pipeline input: False
-Accept wildcard characters: False
-```
-
-### -ResourceGroupName
-Resource Group Name
-
-```yaml
-Type: System.String
-Parameter Sets: ListByParameterSet, GetByNameParameterSet
-Aliases:
-
-Required: True
-Position: 0
-Default value: None
-Accept pipeline input: False
-Accept wildcard characters: False
-```
-
-### -ResourceId
-Azure ResourceId
-
-```yaml
-Type: System.String
-Parameter Sets: GetByResourceIdParameterSet
-Aliases:
-
-Required: True
-Position: Named
-Default value: None
-Accept pipeline input: False
-Accept wildcard characters: False
-```
-
-### CommonParameters
-This cmdlet supports the common parameters: -Debug, -ErrorAction, -ErrorVariable, -InformationAction, -InformationVariable, -OutVariable, -OutBuffer, -PipelineVariable, -Verbose, -WarningAction, and -WarningVariable. For more information, see [about_CommonParameters](http://go.microsoft.com/fwlink/?LinkID=113216).
-
-## INPUTS
-
-### None
-
-## OUTPUTS
-
-### Microsoft.Azure.PowerShell.Cmdlets.DataBoxEdge.Models.PSDataBoxEdgeAlert
-
-## NOTES
-
-## RELATED LINKS
